게시글 작성 서비스 구현

+++ src/main/java/com/example/trainingbygpt/projection/PostProjection.java
@@ -4,5 +4,5 @@ public interface PostProjection {
     Long getPostId();
     String getTitle();
     String getContent();
-    Long getCommentCount();
+    Integer getCommentCount();
 }
